+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master (unreleased)
 
 - Change default threshold from 0.1 to 0
+- Guard against undefined scrollable parent when unmounting
 
 ## 0.1.0